Authorization using policies

+++ app/Http/Controllers/IdeaController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\IdeaRequest;
 use App\Models\Idea;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -45,6 +46,9 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+
+        Gate::authorize('update', $idea);
+
         return view('ideas.show', [
             'idea' => $idea,
         ]);
@@ -55,6 +59,7 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
+        Gate::authorize('create', Idea::class);
         return view('ideas.edit', [
             'idea' => $idea,
         ]);
@@ -65,7 +70,10 @@ class IdeaController extends Controller
      */
     public function update(IdeaRequest $request, Idea $idea)
     {
-        return redirect('/ideas/' . $idea->id);
+        Gate::authorize('create', Idea::class);
+        $idea->update([
+            'description' => request->description,
+        ]);
     }
 
     /**
@@ -73,6 +81,8 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+        Gate::authorize('create', Idea::class);
+
         $idea->delete();
         return redirect('/ideas');
     }
